Refactor RLM eval harness into shared app module and fix eval script imports.
Extract rlm_eval_harness.py so pytest and context_exec_rlm_eval.py share cases
without importing pytest. Fix script sys.path shadowing that broke pydantic.
